requirements.txt
streamlit==1.28.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2 